fix: check-user-status
